// src/types/video.d.ts
declare namespace videojs {
// ...
  namespace AudioTrack {
      /**
       * All possible `AudioTrackKind`s
       *
       * @see https://html.spec.whatwg.org/multipage/embedded-content.html#dom-audiotrack-kind
       */
      type Kind = 'alternative' | 'descriptions' | 'main' | 'main-desc' | 'translation' | 'commentary';
  }
// ...
  type Child =
      | string
      | {
            name: string;
            children?: Child[] | undefined;
        };
// ...
  interface ComponentOptions {
      children?: undefined | Child[];
      createEl?: boolean;
      el?: HTMLElement;
  }
// ...
  interface ControlBarOptions extends ComponentOptions {
      volumePanel?: VolumePanelOptions | boolean | undefined;
      playToggle?: boolean | undefined;
      captionsButton?: boolean | undefined;
      chaptersButton?: boolean | undefined;
      subtitlesButton?: boolean | undefined;
      remainingTimeDisplay?: boolean | undefined;
      progressControl?: ProgressControlOptions | boolean | undefined;
      fullscreenToggle?: boolean | undefined;
      playbackRateMenuButton?: boolean | undefined;
      pictureInPictureToggle?: boolean | undefined;
      currentTimeDisplay?: boolean | undefined;
      timeDivider?: boolean | undefined;
      durationDisplay?: boolean | undefined;
      liveDisplay?: boolean | undefined;
      seekToLive?: boolean | undefined;
      customControlSpacer?: boolean | undefined;
      descriptionsButton?: boolean | undefined;
      subsCapsButton?: boolean | undefined;
      audioTrackButton?: boolean | undefined;
  }
// ...
  interface EventTarget {
      /**
       * An alias of {@link EventTarget#on}. Allows `EventTarget` to mimic
       * the standard DOM API.
       *
       * @param type
       *        An event name or an array of event names.
       *
       * @param fn
       *        The function to call with `EventTarget`s
       *
       * @see {@link EventTarget#on}
       */
      addEventListener(type: string | string[], fn: EventTarget.EventListener): void;

      /**
       * An alias of {@link EventTarget#trigger}. Allows `EventTarget` to mimic
       * the standard DOM API.
       *
       * @param event
       *        The name of the event, an `Event`, or an object with a key of type set to
       *        an event name.
       *
       * @see {@link EventTarget#trigger}
       */
      dispatchEvent(event: string | Event): void;

      /**
       * Removes an `event listener` for a specific event from an instance of `EventTarget`.
       * This makes it so that the `event listener` will no longer get called when the
       * named event happens.
       *
       * @param type
       *        An event name or an array of event names.
       *
       * @param fn
       *        The function to remove.
       */
      off(type: string | string[], fn: EventTarget.EventListener): void;

      /**
       * Adds an `event listener` to an instance of an `EventTarget`. An `event listener` is a
       * function that will get called when an event with a certain name gets triggered.
       *
       * @param type
       *        An event name or an array of event names.
       *
       * @param fn
       *        The function to call with `EventTarget`s
       */
      on(type: string | string[], fn: EventTarget.EventListener): void;

      /**
       * This function will add an `event listener` that gets triggered only once. After the
       * first trigger it will get removed. This is like adding an `event listener`
       * with {@link EventTarget#on} that calls {@link EventTarget#off} on itself.
       *
       * @param type
       *        An event name or an array of event names.
       *
       * @param fn
       *        The function to be called once for each event name.
       */
      one(type: string | string[], fn: EventTarget.EventListener): void;

      /**
       * An alias of {@link EventTarget#off}. Allows `EventTarget` to mimic
       * the standard DOM API.
       *
       * @param type
       *        An event name or an array of event names.
       *
       * @param fn
       *        The function to remove.
       *
       * @see {@link EventTarget#off}
       */
      removeEventListener(type: string | string[], fn: EventTarget.EventListener): void;

      /**
       * This function causes an event to happen. This will then cause any `event listeners`
       * that are waiting for that event, to get called. If there are no `event listeners`
       * for an event then nothing will happen.
       *
       * If the name of the `Event` that is being triggered is in `EventTarget.allowedEvents_`.
       * Trigger will also call the `on` + `uppercaseEventName` function.
       *
       * Example:
       * 'click' is in `EventTarget.allowedEvents_`, so, trigger will attempt to call
       * `onClick` if it exists.
       *
       * @param event
       *        The name of the event, an `Event`, or an object with a key of type set to
       *        an event name.
       */
      trigger(event: string | EventTarget.Event): void;
  }

  const EventTarget: {
      prototype: EventTarget;

      new (): EventTarget;
  };

  namespace EventTarget {
      /**
       * A Custom DOM event.
       * @see [Properties]{@link https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent}
       */
      interface Event extends CustomEvent {
          [key: string]: any;
      }

      /**
       * All event listeners should follow the following format.
       *
       * @param event
       *        the event that triggered this function
       *
       * @param [hash]
       *        hash of data sent during the event
       */
      type EventListener = (e: Event, data?: any) => void;
  }
// ...
  interface KeyboardEvent extends EventTarget.Event {
      readonly which: number;
  }

  interface LanguageTranslations {
      [language: string]: string;
  }
// ...
  interface ModalDialogOptions extends ComponentOptions {
      content?: any;
      description?: string | undefined;
      fillAlways?: boolean | undefined;
      label?: string | undefined;
      temporary?: boolean | undefined;
      uncloseable?: boolean | undefined;
  }
// ...
  interface ProgressControlOptions extends ComponentOptions {
      seekBar?: boolean | undefined;
  }
// ...
  namespace Tech {
      /**
       * An Object containing a structure like: `{src: 'url', type: 'mimetype'}` or string
       * that just contains the src url alone.
       * `var SourceObject = {src: 'http://ex.com/video.mp4', type: 'video/mp4'};`
       * `var SourceString = 'http://example.com/some-video.mp4';`
       */
      interface SourceObject {
          /**
           * The url to the source
           */
          src: string;

          /**
           * The mime type of the source
           */
          type?: string | undefined;
      }
  }
// ...
  interface TextTrackSettingsOptions extends ModalDialogOptions {
      persistTextTrackSettings?: boolean | undefined;
  }
// ...
  interface TextTrack extends Track {
      /**
       * The list text track cues that are currently active for this TextTrack.
       */
      activeCues: TextTrackCueList;

      /**
       * @The text track cue list for this TextTrack.
       */
      cues: TextTrackCueList;

      /**
       * If this track was set to be on or off by default. Cannot be changed after creation.
       */
      readonly default: boolean;

      /**
       * Set the mode of this TextTrack to a valid {@link TextTrack~Mode}. Will
       * not be set if setting to an invalid mode.
       * @fires TextTrack#modechange
       */
      mode: TextTrack.Mode;

      /**
       * Add a cue to the internal list of cues.
       *
       * @param cue
       *        The cue to add to our internal list
       */
      addCue(cue: TextTrackCueList.TextTrackCue): void;

      /**
       * Remove a cue from our internal list
       *
       * @param cue
       *        The cue to remove from our internal list
       */
      removeCue(cue: TextTrackCueList.TextTrackCue): void;
  }
// ...
  interface TextTrackOptions extends TrackOptions {
      tech?: Tech | undefined;
      kind?: TextTrack.Kind | undefined;
      mode?: TextTrack.Mode | undefined;
      srclang?: string | undefined;
      src?: string | undefined;
      default?: boolean | undefined;
  }

  namespace TextTrack {
      type Kind = 'subtitles' | 'captions' | 'descriptions' | 'chapters' | 'metadata';

      type Mode = 'disabled' | 'hidden' | 'showing';
  }
// ...
  interface TrackOptions {
      id?: string | undefined;
      kind?: string | undefined;
      label?: string | undefined;
      language?: string | undefined;
  }
// ...
  interface UserActions {
      click?: boolean | ((event: EventTarget.Event) => void) | undefined;
      doubleClick?: boolean | ((event: EventTarget.Event) => void) | undefined;
      hotkeys?: boolean | ((event: KeyboardEvent) => void) | UserActionHotkeys | undefined;
  }

  interface UserActionHotkeys {
      fullscreenKey?: ((event: KeyboardEvent) => boolean) | undefined;
      muteKey?: ((event: KeyboardEvent) => boolean) | undefined;
      playPauseKey?: ((event: KeyboardEvent) => boolean) | undefined;
  }
// ...
  interface VolumePanelOptions extends ComponentOptions {
      inline?: boolean | undefined;
      volumeControl?: VolumeControlOptions | undefined;
  }
// ...
}
// ...
export interface VideoJsPlayerOptions extends videojs.ComponentOptions {
  aspectRatio?: string | undefined;
  autoplay?: boolean | string | undefined;
  bigPlayButton?: boolean | undefined;
  controlBar?: videojs.ControlBarOptions | false | undefined;
  textTrackSettings?: videojs.TextTrackSettingsOptions | undefined;
  controls?: boolean | undefined;
  defaultVolume?: number | undefined;
  fill?: boolean | undefined;
  fluid?: boolean | undefined;
  height?: number | undefined;
  html5?: any;
  inactivityTimeout?: number | undefined;
  language?: string | undefined;
  languages?: { [code: string]: videojs.LanguageTranslations } | undefined;
  liveui?: boolean | undefined;
  loop?: boolean | undefined;
  muted?: boolean | undefined;
  nativeControlsForTouch?: boolean | undefined;
  notSupportedMessage?: string | undefined;
  playbackRates?: number[] | undefined;
  plugins?: Partial<VideoJsPlayerPluginOptions> | undefined;
  poster?: string | undefined;
  preload?: string | undefined;
  responsive?: boolean | undefined;
  sourceOrder?: boolean | undefined;
  sources?: videojs.Tech.SourceObject[] | undefined;
  src?: string | undefined;
  techOrder?: string[] | undefined;
  tracks?: videojs.TextTrackOptions[] | undefined;
  userActions?: videojs.UserActions | undefined;
  width?: number | undefined;
}

export interface VideoJsPlayerPluginOptions {
  [pluginName: string]: any;
}